Update Assignment 1.java
Assignment 1: Movie Ratings


For this assignment, you will write a program that will aggregate a set of movie ratings for a new film. Sites such as IMDB (Links to an external site.) and RottenTomatoes (Links to an external site.) take lots of data about how much people enjoyed movies and aggregate this to produce a single score indicating how good a movie is. With new movies being released every day it makes sense to use a computer program to calculate these scores.

Your program will compute an overall movie rating as a weighted average of ratings from a popular movie review website, ratings given by a private focus group, and movie critic ratings. You'll see a starter code template for this assignment in your code editor window, and you should write your code in the place indicated.

For input, the program should accept a film title (as a string) and a running time in minutes (as an int). It should then take 3 different integer ratings from a movie review website, then 2 different decimal ratings from a focus group, and finally 1 average rating from professional movie critics which may also be a decimal.

The program should then output the movie name and running time in hours and minutes. This should be followed by the average website rating, the average focus group rating, the average movie critic rating (as doubles), and an overall movie rating (as an int).

The overall movie rating should be computed as a weighted average. Count the average website rating as 20% of the overall rating, the average focus group rating as 30% of the overall rating, and the average movie critic rating as 50% of the overall rating. The overall rating should be rounded to the nearest integer (NOT just truncated to the integer below).

When you write the code to produce your program's final output, take extra care to ensure all the labels used are exactly as shown in the sample run, including the colon after each label. The program which grades your solution will look for the values after these labels, so if they don't match exactly your solution may not be graded correctly.


Milestone 1: Create code that prompts the user for input, taking a String for the movie title and an int for the running time. Write code that displays this title and the number of hours and minutes from the total number of minutes.

Milestone 2: Write code for three int inputs (user ratings), followed by a pair of doubles (focus group ratings) and a single double (critic rating). All of these should be stored in appropriate variables.

Milestone 3: Calculate and display (as doubles) the average of the three user ratings, the average of the two focus group ratings and the single critic rating value. Hint: store the averages calculated as variables so you don’t need to calculate them again later.

Milestone 4: Calculate and display the weighted average of the three different average figures.