Rename Unit 1/Assignment 1.java to Unit 01/Assignment 1.java
Assignment 1: Movie Ratings


For this assignment, you will write a program that will aggregate a set of movie ratings for a new film. Sites such as IMDB and RottenTomatoes take lots of data about how much people enjoyed movies and aggregate this to produce a single score indicating how good a movie is. With new movies being released every day it makes sense to use a computer program to calculate these scores.

Your program will compute an overall movie rating as a weighted average of ratings from a popular movie review website, ratings given by a private focus group, and movie critic ratings. You'll see a starter code template for this assignment in your code editor window, and you should write your code in the place indicated.


Milestone 1: Create code that prompts the user for input, taking a String for the movie title and an int for the running time. Write code that displays this title and the number of hours and minutes from the total number of minutes.

Milestone 2: Write code for three int inputs (user ratings), followed by a pair of doubles (focus group ratings) and a single double (critic rating). All of these should be stored in appropriate variables.

Milestone 3: Calculate and display (as doubles) the average of the three user ratings, the average of the two focus group ratings and the single critic rating value. Hint: store the averages calculated as variables so you don’t need to calculate them again later.

Milestone 4: Calculate and display the weighted average of the three different average figures.