Rename Unit 6/Assignment 6/NumStatsArray.java to Unit 06/Assignment 6/NumStatsArray.java
Assignment 6: Array Statistics


The first class, NumStatsArray, needs to store an array of doubles as a member variable. This should be declared as private and final, so none of your methods will set this variable to point at a different array. The single constructor for this class has a parameter of a double array, and should set the member variable to point to this parameter array.