Rename Unit 6/Assignment 6/runner_StatsArray.java to Unit 06/Assignment 6/runner_StatsArray.java
Assignment 6: Array Statistics


Milestone 1: Add the double array member variable to the NumStatsArrray class and complete the constructor. Write the toString method for the class and test that it formats the output string in the exact format described.

Milestone 2: Write a loop in the average method which sums the values, then use this to return the average. Write a loop to iterate through the array and find the minimum/maximum value in the range method, then use these values to calculate and return the range.

Milestone 3: Write code in the sortStatus method which iterates through the array. Use separate flag variables to determine at the end of the array whether it is in increasing or decreasing order. Add return statements which return the correct values depending on if/how the list is sorted.

Milestone 4: Add the String array member variable to the StringStatsArrray class and complete the constructor. Write the toString method for the class and test that it formats the output string in the exact format described.

Milestone 5: Reuse and modify the code used for the average method in the NumStatsArray class to implement the averageLength method in StringStatsArray. Remember that the lengths of Strings are ints not doubles, but your method must return a double for the average length. Write a simple linear search algorithm in the search method to return the position of the parameter string, or -1 if it does not appear.

Milestone 6: Reuse and modify the code used for the sortStatus method in the NumStatsArray class to write the sortStatus method for the StringStatsArray class. You may need to revise how to test whether one string appears before or after the other alphabetically.