Rename Unit 6/Assignment 6/StringStatsArray.java to Unit 06/Assignment 6/StringStatsArray.java
Assignment 6: Array Statistics

The second class, StringStatsArray, needs to store an array of Strings as a member variable. This should be declared as private and final, so none of your methods will set this variable to point at a different array. The single constructor for this class has a parameter of a String array, and should set the member variable to point to this parameter array.