Rename Unit 7/Assignment 7/runner_GameWheel.java to Unit 07/Assignment 7/runner_GameWheel.java
Assignment 7: Game Wheel

For this assignment, you will work on a multi-file program which represents a game using a spinnable Game Wheel. Your Game Wheel will have 20 evenly cut colored “slices”, each marked with a cash prize amount. Two classes are used to create this wheel: GameWheel and Slice. You will not need to edit the Slice class, but you will implement some of the methods for the GameWheel class. You will also write a method in a third class named Game which can be used to play the game.