Rename Unit 7/Assignment 7/GameWheel.java to Unit 07/Assignment 7/GameWheel.java
Assignment 7: Game Wheel

When a new GameWheel is created it always contains 20 slices, which are represented in the GameWheel class by an ArrayList of Slice objects. The wheel is initialized so that every slice with an odd index is red and every slice with an even index is blue, with the exception of the slices at positions 0, 5, 10 and 15 which are all black. Every slice is initialized with a different prize value.