Rename Unit 9/Assignment 9/runner_Ultimate.java to Unit 09/Assignment 9/runner_Ultimate.java
Assignment 9: Ultimate Frisbee

For this assignment, you will create a hierarchy of five classes to describe various elements of an ultimate frisbee (Links to an external site.) team. Ultimate frisbee is a non-contact sport with players at a position of “cutter” or “handler”. A team usually also has a head coach and possibly one or more assistant coaches. An ultimate team has seven players on the field, with four players at the position of cutter and three at the position of handler. Check out the below video to learn more about this great sport!


Milestone 1: Write the Person class, then write the Coach class extending the Person class with all methods implemented.

Milestone 2: Write the UltimatePlayer class extending the Person class with all methods implemented.

Milestone 3: Write the Captain class extending the UltimatePlayer class with all methods implemented.

Milestone 4: Write the UltimateTeam class and implement all the methods it contains.