Update runner_AnagramList.java
Assignment 10: Anagrams


For this assignment you will write code for a class named AnagramList which is used to find possible anagrams of a word. Anagrams are formed by rearranging the letters of a word: for example brainy is an anagram of binary. For the purpose of this assignment anagrams do not have to be words themselves: for example raibyn would also be an anagram of binary.

The AnagramList class has a single member variable which is an ArrayList of strings called anagrams and an accessor method to get this variable. The constructor takes a single parameter word, and should use this to create an ordered list containing every possible rearrangement of the characters from word exactly once. See the sample runs for some examples of this.


Milestone 1: Initialize the anagrams variable in the constructor and add a call to completeAnagrams with appropriate parameters. Complete the base case for the completeAnagrams method which adds words to the list anagrams. (Hint: think about the simplest possible case for the parameter end.)

Milestone 2: Write code in completeAnagrams for other cases. This will include recursive calls to the same method (completeAnagrams) using different arguments. (Hint: write this initially without worrying about repeats being added to anagrams.)

Milestone 3: Add code to either the constructor or completeAnagrams to ensure that anagrams does not contain any repeats.

Milestone 4: Implement one of the sort algorithms (insertion/selection/merge) in the sortAnagrams method.

Milestone 5: Implement one of the search algorithms (linear/binary) in the searchAnagrams method.