Create Part A
A finite sequence of integers is linear if the differences between all consecutive pairs of elements in the sequence are identical. A finite sequence of integers is quadratic if the differences between all consecutive pairs of elements form a linear sequence. The SequenceProperties class below contains a variety of static methods which are used to interpret and analyze sequences.