Create Coding Activity 1.java
Write a program to print the name LeBron James to the center of the output screen.

To center your output, you should use spaces in your String (at the beginning and end). The number of spaces on either side should depend on how wide your screen is.

For grading purposes, only one space will be needed on both sides of the name for your program to be counted as centered and acceptable.