Create Coding Activity 2.java
Write a program which prints the name Alicia Keys three times. Each time the program prints the name it should be on a new line and in the center of the output screen. To pass the tests, your output must have at least one space before and after Alicia Keys' name, and have the same number of spaces before and after the name on each line. See the sample run below for an example of this.