Create Coding Activity 1.java
Write code to ask the user to enter their name and print the following message:

Hi ______, nice to see you.

Where ______ is replaced with the name input by the user. Remember, you'll need to create a Scanner and use the method nextLine on it to get the user input and store it in a variable.