Create Coding Activity 2.java
Write a program that asks the user their name and age and prints out this information in a sentence as shown. Note that age should be an integer, not a double.

Sample run:
Hi there. What is your name?
Pascal
Hi Pascal. How old are you?
16
Pascal is 16 years old.