Create Coding Activity 3.java
Input the radius of a circle and print the circumference and area. The answers should be decimal values and must be printed in the format shown in the sample run.

Remember the formula for circumference is 2 * pi * r and the formula for area is pi * r * r. You can use 3.14 for the value of pi in this activity.

Sample Run:
Enter a radius:
4
Circumference: 25.12
Area: 50.24