Create Coding Activity 4.java
Write code which asks the user to enter the price of an item and tells them how much money they would get back from $20. It's okay for this to be negative. Your output must be in the exact format shown, including the dollar signs.

Sample Run:
Enter a price:
3.50
Change from $20: $16.5