Create Coding Activity 1.java
Try to run the following code:
double x = (11 + 4) / 2;
System.out.println(x);

What answer do you get? The answer is imprecise!

Write a Java program that precisely calculates the average between two integers that input from the keyboard, and prints the answer to the screen. The average displayed must be accurate, including when the average of the two numbers is a decimal.


Sample Run:
Please enter two integers:
4
11

The average is: 7.5