Create Coding Activity 1.java
Write a program which asks the user for their name and age. The program should then print a sentence repeating this information as shown in the sample run.

Sample run:
What is your name?
Cory
What is your age?
48
Cory is 48 years old.

Note: All the user input prompts need to match the sample run exactly to get full credit for this activity.