Create Coding Activity 1.java
Write a program that asks the user for an item type to buy, a number of items and a cost (remember to think about what data type should be used for each one). The program should then print a sentence repeating this information in a sentence as shown in the sample run.

Sample run:
What type of item are you buying?
widgets
How many are you buying?
11
How much does each one cost?
3.75
11 widgets at $3.75 each will cost you $41.25

Note: All the user input prompts need to match the sample run exactly to get full credit for this activity.