Create Coding Activity 1.java
Write code to take a String input from the user, then print the first and last letters of the string on one line.

Sample run:
Enter a string:
surcharge
se

Hint - we saw in the lesson how to get the first letter of a string using the substring method. Think how you could use the .length() String method to find the index of the last character, and how you could use this index in the substring method to get the last character.