Create Coding Activity 1.java
In this exercise you will debug the code which has been provided in the starter file. The code is intended to take two strings, s1 and s2 followed by a whole number, n, as inputs from the user, then print a string made up of the first n letters of s1 and the last n letters of s2. Your job is to fix the errors in the code so it performs as expected (see the sample run for an example).

Sample run
Enter first string
sausage
Enter second string
races
Enter number of letters from each word
3
sauces

Note: you are not expected to make your code work when n is bigger than the length of either string.