Create Coding Activity 2.java
Write code which takes a double input from the user and creates a circle with that radius. The code should calculate the area of the circle, then double the radius of the circle and print the new area of the circle.

Sample run:
Enter radius:
> 2.0
Area: 12.566370614359172
Radius is being doubled...
Area: 50.26548245743669

Hint: You should use the getRadius and setRadius methods to double the radius.