Create Coding Activity 3.java
Write a program to test if a double input by the user is not a valid test score. The program should print "Not Valid" (not including the quotes) only if the number entered by the user is NOT between 0 and 100 inclusive.

Sample run 1:
Enter a test score:
75.5

Sample run 2:
Enter a test score:
102.5
Not Valid

Hint - you will need more than one if statement to be sure that the score is not between 0 and 100.