Create Coding Activity 4.java
You are running an experiment that involves hatching chicken eggs. Bird eggs are very sensitive to temperature and chickens’ eggs will hatch between 99 and 102 degrees Fahrenheit.

Write the code for the sensor that will be tracking the temperature. If the temperature falls below 99 or above 102 your code should print “WARNING”. Your code should accept the temperature input as a double.

Sample run 1:
What is the temperature?
45.3
WARNING

Sample run 2:
What is the temperature?
100.0