Create Coding Activity 2.java
Write a program that takes input as an integer grade from the keyboard and translates it to a letter grade. For example, if a user enters 4, print "A", and if they enter 0 print "F."  You can assume that an integer between 0 and 4 will be input. To gain full credit for this exercise you will need to use else if statements when checking the grade.

Note - in this grading system there is no grade "E". Therefore the numbers and grades are linked as follows: A-4, B-3, C-2, D-1, F-0.