Create Coding Activity 3.java
Create a program to let the user practice their multiplication tables. Print a multiplication question using two random integers between 1 and 12 as shown in the sample run. Then allow the user input an int as their answer to the multiplication of the two numbers. If they enter the correct product print “Correct!” otherwise print “Wrong”. You must use an else statement in your code to receive full credit for this exercise.

Sample run 1:
4 * 7 = ?
27
Wrong

Sample run 2:
8 * 2 = ?
16
Correct!

Note - the code for this exercise imports a class named edhesive.testing.Math. This is a class used to test and grade your code. Please do not remove this import statement as your code will then not be graded correctly.