Create Coding Activity 1.java
The template code provided is intended to check whether an integer entered by the user is outside of the range 20-29 (inclusive). If it is outside of this range the program should print a warning and change the number to 25. However, when using De Morgan's law to simplify this code, the programmer has made some mistakes. Can you correct the errors so the code functions as intended?

Sample runs of the program functioning properly are below:

Sample run 1:
Enter a number in the twenties
28
Your number is 28

Sample run 2:
Enter a number in the twenties
33
That's not in the twenties!
Your number is 25