Create Coding Activity 2.java
The template code provided is intended to take two inputs, x and y, from the user and print "pass" if one or more of the following is true:

• x is not less than 4
• y is not greater than 5 and x + y is less than 7

However, when using De Morgan's law to simplify this code, the programmer has made some mistakes. Can you correct the errors so the code functions as intended?