Create Coding Activity 2.java
Write a program that asks the user to enter 2 words, then prints "Great!" if the two words are identical, "Close enough" if they are the same length and all but the last letter matches, and "Try again" otherwise.

Sample runs:
Enter 2 strings:
the
the
Great!

Enter 2 strings:
that
than
Close enough

Enter 2 strings:
their
there
Try again

Hint - for the "Close enough" condition you will need to see whether substrings of the two input String are the same. You can use the length and substring methods together to get the relevant substrings (all but the last character of each String).