Create Coding Activity 3.java
Input an int between 0 and 100 and print the numbers between it and 100, including the number itself and the number 100. If the number is less than or equal to 0, or greater than or equal to 100 print "error". Print 20 numbers per line.

Hint - use modular division to determine where you need to add a line break.

Sample Run 1:
Enter a number between 0 and 100:
30
30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49
50 51 52 53 54 55 56 57 58 59 60 61 62 63 64 65 66 67 68 69
70 71 72 73 74 75 76 77 78 79 80 81 82 83 84 85 86 87 88 89
90 91 92 93 94 95 96 97 98 99 100

Sample Run 2:
Enter a number between 0 and 100:
105
error