Create Coding Activity 1.java
Write a method that takes a parameter for the number of a month and prints the month's name. You may assume that the actual parameter value passed to the method is always between 1 and 12 inclusive.

This method must be called monthName() and it must have an integer parameter.

Calling monthName(8) should print August to the screen.

You can call your method in the program's main method so you can test whether it works, but you must remove or comment out the main method before checking your code for a score.