Create Coding Activity 4.java
Write a method that accepts a number of seconds and prints the correct number of hours, minutes, and seconds.

This method must be called realTime() and its parameter must be an integer.

Calling realTime(6342) would print the following:
Hours: 1
Minutes: 45
Seconds: 42

You can call your method in the program's main method so you can test whether it works, but you must remove or comment out the main method before checking your code for a score.

Hint - you might want to review the unit 1 lessons on the use of integer and modular division as you will need to use these in your answer. Start by thinking about how many seconds there are in an hour, and therefore how you would work out the number of whole hours in a given number of seconds (and how many seconds would be leftover).