Create Coding Activity 2.java
Write a method that takes 5 ints as parameters and returns the average value of the 5 ints as a double.

This method must be named average() and it must have 5 int parameters. This method must return a double.

Calling average(1, 5, 7, 4, 10) would return 5.4.

You can call your method in the program's main method so you can test whether it works, but you must remove or comment out the main method before checking your code for a score.