Create Coding Activity 4.java
Write a method that takes a String parameter. If the String has a double letter (i.e. contains the same letter twice in a row) then it should return true. Otherwise, it should return false.

This method must be named hasRepeat() and have a String parameter. This method must return a boolean.

You can call your method in the program's main method so you can test whether it works, but you must remove or comment out the main method before checking your code for a score.