Create Post.java
Create a class named Post which will represent a social media post. The class should store the following information in appropriate (private) variables:

The name of the user who wrote the post
The (text) contents of the post
The number of people who have “favorited” the post
You should write a constructor for the class with the signature Post(String user, String contents, int favorites), which sets the three variables for the class. You should also write a toString method that formats the Post in the following way.

name
Contents of message
    	Favorited by favorites users

Note that the third line of the string starts with a single tab character.

You'll need to add member variables (fields) to your class to store the information for a post - make sure you choose the names of these carefully to avoid a conflict with the names of the parameters you use in your constructor.