Create runner_Elevator.java
To test your code before submission, use the file runner_Elevator.java. Run the main method in the runner_Elevator class and verify that the program produces the expected output. A sample run is shown below, but you should also try a range of other inputs to test your code more thoroughly.

If the runner file does not compile then you should attempt to fix your Elevator class before checking your code for a score. Start by checking that the signatures and return types of your Elevator methods match the descriptions exactly. You should also not add a main method to your Elevator class, since you may not be able to run your code, and it will not be scored correctly when you check it.

Sample run:
How many floors?
5
Starting on which floor?
2
Elevator is on floor 2 of 5.
Elevator in service.
Type "c" to change floor, "s" to put elevator in or out of service, "r" to restart, "q" to quit.
c
New floor number?
4
Elevator is on floor 4 of 5.
Elevator in service.
Type "c" to change floor, "s" to put elevator in or out of service, "r" to restart, "q" to quit.
s
Elevator is on floor 4 of 5.
Elevator not in service.
Type "c" to change floor, "s" to put elevator in or out of service, "r" to restart, "q" to quit.
q

Hint - Start by writing the 3 accessor methods - getMaxFloor, getCurrentFloor and isInService. These methods just need to return the relevant member variables. Once this is done you will be able to compile and run the runner class to test your code. Then you can focus on writing and testing the more difficult remaining methods.