Create Student.java
For this exercise, you will create a Student class to hold student data. You should create appropriate variables in your Student class to store the following information:

• First name
• Last name
• Grade level - an integer value ranging from 0 to 12, with 0 representing kindergarten.
• GPA - a decimal value ranging from 0 to 4.5.
• Student ID - an integer value which is automatically assigned (see below).

The student IDs are assigned in order: the first student has an ID of 1, the second student ID is 2, the next is 3, and so on. When a new Student object is created, your code should automatically initialize student id to the next integer value. For example, if the last created Student's student id is 17, the next student id generated should be 18.