Create U6_L2_Activity_Two.java
In the U6_L2_Activity_Two class, write a public static method called numEven, which has a single parameter of an array of int values. The method should return the number of elements in the array which are even. Use the runner class to test this method: do not add a main method to your code in the U6_L2_Activity_Two.java file or it will not be scored correctly.

Sample runs (runner class):
Enter array length:
4
Enter values:
4 2 7 10
Number even: 3

Enter array length:
6
Enter values:
1 34 7 12 23 49
Number even: 2