Create runner_U6_L3_Activity_Two.java
Use the runner class to test this method: do not add a main method to your code in the U6_L3_Activity_Two.java file or it will not be scored correctly.

Hint: You might want to look back at Unit 4: Lesson 5 - Coding Activity 1, which also required you to print the letters of a String in reverse order. For this activity though you will need your outer loop to iterate through the array, and your inner loop to iterate through the string.